requirements.txt
django==2.0.2
django-livereload-server
requests
pytest-django
